OutcomeIntegrationTests & supporting improvements
* NSDate is now frozen for all tests
   - This is to prepare for on_focus outcome tests
   - Updated tests that were effected by this change.
* Added OutcomeIntegrationTests.m
   - Just a start on test
   - More to come, just a starting point file for now.
   - Will be adding the on_focus tests here in another PR.
* Added mocking of outcomes for ios_params get call
* Added RestClientAsserts like we have on Android.
* Added some helpers to UnitTestCommonMethods

+++ iOS_SDK/OneSignalSDK/Source/OneSignalTracker.m
@@ -186,7 +186,7 @@ static BOOL lastOnFocusWasToBackground = YES;
     let now = [[NSDate date] timeIntervalSince1970];
     [OneSignalSharedUserDefaults saveDouble:now withKey:USER_LAST_CLOSED_TIME];
    
-    let timeElapsed = now - lastOpenedTime + 0.5;
+    let timeElapsed = now - (int)(lastOpenedTime + 0.5);
    
     // Time is invalid if below 1 or over a day
     if (timeElapsed < 0 || timeElapsed > 86400)
